fix(Upload): BC - Fix detection of finished uploads
BREAKING CHANGE:
- for some frontend the last chunk request must contain the info about file -> let's wait for all the chunks
- as this is a breaking change to prevent issue on upload file name is always hashed

+++ _setup/setupVersion.js
@@ -7,6 +7,7 @@ const getLaravelsRoutesPath = require('./getLaravelsRoutesPath.js');
 const modifyComposer = require('./composer/modifyComposer.js');
 const addServiceProvider = require('./addServiceProvider.js');
 const publish = require('./publish.js');
+const updateEnvForDockerLogs = require('./updateEnvForDockerLogs.js');
 
 /**
  * @param {{laravel: string, image_version: string}} version
@@ -14,7 +15,12 @@ const publish = require('./publish.js');
  * @param {boolean} verbose
  * @param {number} index
  */
-async function setupVersion (version, currentDirectory, verbose, index = 0) {
+async function setupVersion (
+    version,
+    currentDirectory,
+    verbose,
+    publishOnlyExample = false,
+) {
     const numericLaravelVersion = parseFloat(
         version.laravel === 'dev-master' ? 999 : version.laravel.replace('.*', '')
     );
@@ -25,57 +31,60 @@ async function setupVersion (version, currentDirectory, verbose, index = 0) {
     console.log('');
 
     if (verbose) {
-        console.log('Directory: ',currentDirectory);
-        console.log('Version directory: ',versionDir);
+        console.log('Directory: ', currentDirectory);
+        console.log('Version directory: ', versionDir);
     }
 
-    const envComposer = {
-        'COMPOSER_NO_SECURITY_BLOCKING': 1, // due a support of older version,
-        'COMPOSER_NO_BLOCKING': 1
-    }
-
-    // Update or install Laravel project
-    if (fs.existsSync(path.resolve(versionDir, 'composer.json'))) {
-        await run('composer', ['update'], versionDir, verbose, envComposer);
-    } else {
-        if (fs.existsSync(versionDir)) {
-            fs.rmdirSync(versionDir)
+    if (publishOnlyExample === false) {
+        const envComposer = {
+            'COMPOSER_NO_SECURITY_BLOCKING': 1, // due a support of older version,
+            'COMPOSER_NO_BLOCKING': 1
         }
 
-        const options = [
-            'create-project',
-            '--prefer-dist',
-            `laravel/laravel`,
-            version.laravel,
-            version.laravel,
-        ];
-        await run('composer', options, currentDirectory, verbose, envComposer);
+        // Update or install Laravel project
+        if (fs.existsSync(path.resolve(versionDir, 'composer.json'))) {
+            await run('composer', ['update'], versionDir, verbose, envComposer);
+        } else {
+            if (fs.existsSync(versionDir)) {
+                fs.rmdirSync(versionDir)
+            }
+
+            const options = [
+                'create-project',
+                '--prefer-dist',
+                `laravel/laravel`,
+                version.laravel,
+                version.laravel,
+            ];
+            await run('composer', options, currentDirectory, verbose, envComposer);
+        }
+
+        // Delete default routes path and create a an empty php file.
+        const routesPath = getLaravelsRoutesPath(numericLaravelVersion, versionDir)
+
+        updateEnvForDockerLogs(versionDir);
+
+        await deleteIfExists(routesPath, currentDirectory);
+
+        fs.writeFileSync(routesPath, '<?php\n', 'utf8');
+
+        console.log(colors.green('⛑  Altering composer.json to support local example'))
+        modifyComposer(versionDir)
+
+        const requireExampleArgs = ['require', 'pion/laravel-chunk-example-app'];
+        await run('composer', requireExampleArgs, versionDir, verbose)
+
+        // For older Laravel versions without auto-discovery
+        if (numericLaravelVersion < 5.5) {
+            console.log(colors.green('⛑  Altering app config to use example service provider'))
+
+            const exampleServiceProvider = 'Pion\\Laravel\\ChunkUploadExample\\ExampleServiceProvider::class';
+            addServiceProvider(path.resolve(versionDir, 'config', 'app.php'), exampleServiceProvider);
+        }
+
+        await run('php', ['artisan', 'storage:link'], versionDir, verbose);
     }
-
-    // Delete default routes path and create a an empty php file.
-    const routesPath = getLaravelsRoutesPath(numericLaravelVersion, versionDir)
-
-    await deleteIfExists(routesPath, currentDirectory);
-
-    fs.writeFileSync(routesPath, '<?php\n', 'utf8');
-
-    console.log(colors.green('⛑  Altering composer.json to support local example'))
-    modifyComposer(versionDir)
-
-    const requireExampleArgs = ['require', 'pion/laravel-chunk-example-app'];
-    await run('composer', requireExampleArgs, versionDir, verbose)
-
-    // For older Laravel versions without auto-discovery
-    if (numericLaravelVersion < 5.5) {
-        console.log(colors.green('⛑  Altering app config to use example service provider'))
-
-        const exampleServiceProvider = 'Pion\\Laravel\\ChunkUploadExample\\ExampleServiceProvider::class';
-        addServiceProvider(path.resolve(versionDir, 'config', 'app.php'), exampleServiceProvider);
-    }
-
     await publish(versionDir, verbose);
-
-    await run('php', ['artisan', 'storage:link'], versionDir, verbose);
 }
 
 module.exports = setupVersion
